// pod_automation_frontend/src/api/models/ListingOptimizationRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ListingOptimizationRequest = {
    listing_id: string;
    current_title?: string;
    current_tags?: Array<string>;
    current_description?: string;
};